// gauvendi-platform-service-v2/src/core/enums/common.ts
export enum RatePlanPaymentSettlementSettingModeEnum {
  PMS_SETTLEMENT = 'PMS_SETTLEMENT',
  MANUAL_SETTLEMENT = 'MANUAL_SETTLEMENT'
}
// ...
export enum ResponseCodeEnum {
  SUCCESS = '200',
  PENDING = '202',
  ERROR = '400'
}