// gauvendi-platform-service-v2/src/core/enums/booking-transaction.ts
export enum BookingTransactionStatusEnum {
  PAYMENT_PENDING = 'PENDING_PAYMENT',
  PAYMENT_SUCCEEDED = 'PAYMENT_SUCCEEDED',
  PAYMENT_FAILED = 'PAYMENT_FAILED'
}

export enum ProcessPaymentValidationMessage {
  NOT_FOUND_PAYMENT_PROVIDER = 'NOT_FOUND_PAYMENT_PROVIDER',
  NOT_FOUND_PROPERTY_PAYMENT_METHOD_SETTING = 'NOT_FOUND_PROPERTY_PAYMENT_METHOD_SETTING',
  NOT_FOUND_PROPERTY_PAYMENT_METHOD_SETTING_METADATA = 'NOT_FOUND_PROPERTY_PAYMENT_METHOD_SETTING_METADATA',
  NOT_FOUND_PAYMENT_API_KEY = 'NOT_FOUND_PAYMENT_API_KEY',
  SHOPPER_REFERENCE_REQUIRED = 'SHOPPER_REFERENCE_REQUIRED',
  REQUEST_BOOKING_PAYMENT_FAILED = 'REQUEST_BOOKING_PAYMENT_FAILED',
  NOT_FOUND_PAYMENT_DATA = 'NOT_FOUND_PAYMENT_DATA',
  REQUIRE_ACTION = 'REQUIRE_ACTION',
  PROCESS_PAYMENT_SUCCESS = 'PROCESS_PAYMENT_SUCCESS',
  REQUIRE_CARD_INFORMATION = 'REQUIRE_CARD_INFORMATION',
  NOT_FOUND_REF_PAYMENT_METHOD_ID = 'NOT_FOUND_REF_PAYMENT_METHOD_ID'
}

export enum ResponseContentStatusEnum {
  ERROR = 'ERROR',
  PENDING = 'PENDING',
  SUCCESS = 'SUCCESS'
}

export enum BookingValidationMessage {
  INVALID_INPUT = 'INVALID_INPUT',
  NOT_FOUND_BOOKING = 'NOT_FOUND_BOOKING',
  NOT_FOUND_PAYMENT_INTENT = 'NOT_FOUND_PAYMENT_INTENT',
  BOOKING_PAYMENT_NOT_YET_COMPLETED = 'BOOKING_PAYMENT_NOT_YET_COMPLETED',
  COMPLETE_BOOKING_SUCCESS = 'COMPLETE_BOOKING_SUCCESS'
}